Refactor: seperate types from components to types/
